// src/generated/pistonpanel/instance.ts
export interface InstanceUser {
  /**
   * @generated from protobuf field: string id = 1;
   */
  id: string;
  /**
   * @generated from protobuf field: string username = 2;
   */
  username: string;
  /**
   * @generated from protobuf field: string email = 3;
   */
  email: string;
}
// ...
export interface InstanceAuditLogResponse {
  /**
   * @generated from protobuf field: repeated pistonpanel.v1.InstanceAuditLogResponse.AuditLogEntry entry = 1;
   */
  entry: InstanceAuditLogResponse_AuditLogEntry[];
}
/**
 * @generated from protobuf message pistonpanel.v1.InstanceAuditLogResponse.AuditLogEntry
 */
export interface InstanceAuditLogResponse_AuditLogEntry {
  /**
   * @generated from protobuf field: string id = 1;
   */
  id: string;
  /**
   * @generated from protobuf field: pistonpanel.v1.InstanceUser user = 2;
   */
  user?: InstanceUser;
  /**
   * @generated from protobuf field: pistonpanel.v1.InstanceAuditLogResponse.AuditLogEntryType type = 3;
   */
  type: InstanceAuditLogResponse_AuditLogEntryType;
  /**
   * @generated from protobuf field: google.protobuf.Timestamp timestamp = 4;
   */
  timestamp?: Timestamp;
  /**
   * @generated from protobuf field: string data = 5;
   */
  data: string;
}
/**
 * @generated from protobuf enum pistonpanel.v1.InstanceAuditLogResponse.AuditLogEntryType
 */
export enum InstanceAuditLogResponse_AuditLogEntryType {
  /**
   * @generated from protobuf enum value: EXECUTE_COMMAND = 0;
   */
  EXECUTE_COMMAND = 0,
  /**
   * @generated from protobuf enum value: START_ATTACK = 1;
   */
  START_ATTACK = 1,
  /**
   * @generated from protobuf enum value: PAUSE_ATTACK = 2;
   */
  PAUSE_ATTACK = 2,
  /**
   * @generated from protobuf enum value: RESUME_ATTACK = 3;
   */
  RESUME_ATTACK = 3,
  /**
   * @generated from protobuf enum value: STOP_ATTACK = 4;
   */
  STOP_ATTACK = 4,
}
// ...
class InstanceUser$Type extends MessageType<InstanceUser> {
  constructor() {
    super('pistonpanel.v1.InstanceUser', [
      { no: 1, name: 'id', kind: 'scalar', T: 9 /*ScalarType.STRING*/ },
      { no: 2, name: 'username', kind: 'scalar', T: 9 /*ScalarType.STRING*/ },
      { no: 3, name: 'email', kind: 'scalar', T: 9 /*ScalarType.STRING*/ },
    ]);
  }
}
/**
 * @generated MessageType for protobuf message pistonpanel.v1.InstanceUser
 */
export const InstanceUser = new InstanceUser$Type();
// ...
export const InstanceAuditLogResponse = new InstanceAuditLogResponse$Type();
// @generated message type with reflection information, may provide speed optimized methods
class InstanceAuditLogResponse_AuditLogEntry$Type extends MessageType<InstanceAuditLogResponse_AuditLogEntry> {
  constructor() {
    super('pistonpanel.v1.InstanceAuditLogResponse.AuditLogEntry', [
      { no: 1, name: 'id', kind: 'scalar', T: 9 /*ScalarType.STRING*/ },
      { no: 2, name: 'user', kind: 'message', T: () => InstanceUser },
      {
        no: 3,
        name: 'type',
        kind: 'enum',
        T: () => [
          'pistonpanel.v1.InstanceAuditLogResponse.AuditLogEntryType',
          InstanceAuditLogResponse_AuditLogEntryType,
        ],
      },
      { no: 4, name: 'timestamp', kind: 'message', T: () => Timestamp },
      { no: 5, name: 'data', kind: 'scalar', T: 9 /*ScalarType.STRING*/ },
    ]);
  }
}
/**
 * @generated MessageType for protobuf message pistonpanel.v1.InstanceAuditLogResponse.AuditLogEntry
 */
export const InstanceAuditLogResponse_AuditLogEntry =
  new InstanceAuditLogResponse_AuditLogEntry$Type();